HW-4 BasePresenter 생성 및 Activity  적용
