+++ toctocbin/AndroidArchitecture/app/src/main/java/com/test/androidarchitecture/base/BasePresenter.kt
@@ -1,0 +1,12 @@
+package com.test.androidarchitecture.base
+
+import io.reactivex.disposables.CompositeDisposable
+
+abstract class BasePresenter : BaseContract.Presenter {
+
+    protected val compositeDisposable = CompositeDisposable()
+
+    override fun clearDisposable() {
+        compositeDisposable.clear()
+    }
+}
